requirements.txt
aniso8601==9.0.1
attrs==21.4.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
coverage==6.3.2
Faker==13.6.0
Flask==2.1.1
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
flask-mongoengine==1.0.0

flask-restx==0.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
iniconfig==1.1.1
itsdangerous==2.1.2
jaeger-client==4.8.0
Jinja2==3.1.1
jsonschema==4.4.0
MarkupSafe==2.1.1
mongoengine==0.24.1
opentracing==2.4.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyJWT==1.7.1
pymongo==4.1.1
PyMySQL==1.0.2
pyparsing==3.0.8
pyrsistent==0.18.1
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.35
threadloop==1.0.2
thrift==0.16.0
tomli==2.0.1
tornado==6.1
urllib3==1.26.9
Werkzeug==2.1.1
WTForms==3.0.1
zipp==3.8.0
